feat(plugin-inquirer): add figures symbols
